+++ datumaro/util/definitions.py
@@ -6,3 +6,4 @@ from typing import Tuple
 
 DEFAULT_SUBSET_NAME = "default"
 BboxIntCoords = Tuple[int, int, int, int]  # (x, y, w, h)
+SUBSET_NAME_BLACKLIST = {"labels", "images", "annotations", "instances"}