+++ tests/unit/test_vgg_face2_format.py
@@ -283,7 +283,7 @@ class VggFace2ImporterTest(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_detect(self):
         detected_formats = Environment().detect_dataset(DUMMY_DATASET_DIR)
-        self.assertIn(VggFace2Importer.NAME, detected_formats)
+        self.assertEqual([VggFace2Importer.NAME], detected_formats)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     def test_can_import(self):