+++ CHANGELOG.md
@@ -47,6 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/854>)
 - Fix mask visualization bug
   (<https://github.com/openvinotoolkit/datumaro/pull/860>)
+- Fix detect unit tests to test false negatives as well
+  (<https://github.com/openvinotoolkit/datumaro/pull/868>)
 
 ## 24/02/2023 - Release v1.0.0
 ### New features